Fix spelling in category names (balance sheet/GuV).
I was unsure if this is safe, since these category names are hashed and
basis for the cid which is associated with each ledger account.

But the string that is hashes it the "matched category", which is the
category description as is used in skr03-ocr-results.json (so how the
category was named in the SKR03 PDF).

parse_chart_of_accounts.rb creates then a fuzzy match, and the json
files contain the "official" category name (from HGB) and the hashed
string of how the category is named in the PDF.

Because of this, the hash stays the same, I only tweaked the official
category name.